refactor : 파일 삭제, 디렉토리 이동
