+++ src/main/java/honbob/honbob/dto/kakao/KakaoLoginRequest.java
@@ -1,4 +1,4 @@
-package honbob.honbob.dto;
+package honbob.honbob.dto.kakao;
 
 import lombok.Getter;
 import lombok.Setter;
