+++ src/main/java/honbob/honbob/dto/kakao/KakaoUserInfoDto.java
@@ -1,4 +1,4 @@
-package honbob.honbob.dto;
+package honbob.honbob.dto.kakao;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
