--- src/test/java/honbob/honbob/HonbobApplicationTests.java
@@ -1,13 +1,0 @@
-package honbob.honbob;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class HonbobApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
